Moved DB.Face extension methods to External.DB.Extensions namespace.
