+++ src/RhinoInside.Revit/RhinoInside.Revit.csproj
@@ -171,6 +171,7 @@
     <Compile Include="Exceptions.cs" />
     <Compile Include="Extensions\RevitAPI.cs" />
     <Compile Include="External\DB\Extensions\Curve.cs" />
+    <Compile Include="External\DB\Extensions\Face.cs" />
     <Compile Include="External\DB\Extensions\XYZ.cs" />
     <Compile Include="External\DB\UniqueId.cs" />
     <Compile Include="External\UI\Application.cs" />
